+++ kernel/libkernelcxx/CMakeLists.txt
@@ -4,9 +4,11 @@ set(libkernelcxx_HEADERS
 	include/algorithm.hpp
 	include/array.hpp
 	include/atomic.hpp
+	include/concepts.hpp
 	include/cstddef.hpp
 	include/cstdint.hpp
 	include/cstring.hpp
+	include/cxxabi.h
 	include/ext/fixed_size_vector.hpp
 	include/ext/intrusive_circular_list.hpp
 	include/ext/memory.hpp
@@ -23,27 +25,23 @@ set(libkernelcxx_HEADERS
 	include/impl/type_traits/make_unsigned.hpp
 	include/impl/type_traits/remove_cv.hpp
 	include/memory.hpp
+	include/new.hpp
 	include/ostream.hpp
 	include/streambuf.hpp
+	include/string.hpp
 	include/type_traits.hpp
 	include/utility.hpp
-
-	src/cstring.cpp
-	src/ext/scoped_format_guard.cpp
-	src/ext/snprintf_stream.cpp
-	src/ostream.cpp
-	src/streambuf.cpp
 )
 
 set(libkernelcxx_SOURCES
 	src/cstring.cpp
+	src/cxxabi.cpp
+	src/ext/scoped_format_guard.cpp
+	src/ext/snprintf_stream.cpp
+	src/new.cpp
 	src/ostream.cpp
 	src/streambuf.cpp
 	src/string.cpp
-	src/new.cpp
-
-	src/ext/scoped_format_guard.cpp
-	src/ext/snprintf_stream.cpp
 )
 
 add_library(libkernelcxx STATIC
@@ -52,4 +50,4 @@ add_library(libkernelcxx STATIC
 )
 
 target_include_directories(libkernelcxx PUBLIC include/)
-target_link_libraries(libkernelcxx PUBLIC log)
+target_link_libraries(libkernelcxx PUBLIC log memory)
