+++ kernel/main.cpp
@@ -1,5 +1,5 @@
-#include <bolgenos-ng/cxxabi.h>
 
+#include <cxxabi.h>
 #include <bolgenos-ng/asm.hpp>
 #include <bolgenos-ng/interrupt_controller.hpp>
 #include <bolgenos-ng/irq.hpp>
